src/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agent-framework-azure-ai>=1.0.0b251007",
    "agent-framework-core>=1.0.0b251007",
    "azure-ai-agents>=1.2.0b5",
    "azure-ai-projects>=1.1.0b4",
    "python-dotenv>=1.2.1",
]
